requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
playwright
beautifulsoup4
python-dotenv
markdownify
pydantic
Pillow
langchain-chroma
dotenv
urllib3
beautifulsoup4
requests


